Code cleanup to make part 2 easier
Use smaller data types for player fields, just in case we need
to make a lot of them.

+++ src/main/kotlin/day21/Day21.kt
@@ -56,17 +56,17 @@ internal class DeterministicDie {
 internal fun parse(input: List<String>): LinkedList<Player> {
     val players = input
         .map { it.toCharArray() }
-        .map { Pair(it[7].toString().toInt() , it.last().toString().toInt() - 1) }
+        .map { Pair(it[7].toString().toByte() , (it.last().toString().toInt() - 1).toByte()) }
         .map { Player(it.first, it.second, 0) }
         .toList()
 
     return LinkedList(players)
 }
 
-internal data class Player(val id: Int, val position: Int, val score: Int) {
+internal data class Player(val id: Byte, val position: Byte, val score: Short) {
     fun advance(spaces: Int) : Player {
         val newPosition = (position + spaces) % 10
         val newScore = score + newPosition + 1
-        return Player(id, newPosition, newScore)
+        return Player(id, newPosition.toByte(), newScore.toShort())
     }
 }
